L6 Commit 4: Add /analyze route to process form input and render results

+++ main.py
@@ -26,7 +26,7 @@ def collect_numbers():
         while True:
             s = input("Enter an integer: ").strip()
             try:
-                num = int(s)
+                num = float(s)
                 numbers.append(num)
                 break
             except ValueError:
